+++ imap_tools/__init__.py
@@ -7,4 +7,4 @@ from .errors import *
 
 __author__ = 'Vladimir Kaukin <KaukinVK@ya.ru>'
 
-__version__ = '0.44.0'
+__version__ = '0.45.0'