+++ imap_tools/folder.py
@@ -77,7 +77,7 @@ class MailBoxFolderManager:
         """
         Get a listing of folders on the server
         :param folder: mailbox folder, if empty - get from root
-        :param search_args: search argumets, is case-sensitive mailbox name with possible wildcards
+        :param search_args: search arguments, is case-sensitive mailbox name with possible wildcards
             * is a wildcard, and matches zero or more characters at this position
             % is similar to * but it does not match a hierarchy delimiter
         :param subscribed_only: bool - get only subscribed folders